==> Final R Scripts/4_Statistical Analysis.Rmd <==
---
title: "R Notebook"
output: html_notebook
editor_options: 
  chunk_output_type: console
---


```{r}
library(tidyverse)
library(psych)
library(ez)
library(export)
```


```{r}
#summarised_data <- read_csv(file.choose())
summarised_driving_data <- read_csv(file = "G:\\Shared drives\\Projects\\2018 Driving Sim Reproducibility\\summarised_DB.csv") %>% select(-1)
```

```{r}

create_ICC_dataframe_driving <- function(database,variable,data_name){
    data <- database %>% select(ID,task,session,run,variable)
    U_F1    <- data %>% filter(task == "Urban" & session == 0 & run == 1)
    U_S1_R1 <- data %>% filter(task == "Urban" & session == 1 & run == 1)
    U_S1_R2 <- data %>% filter(task == "Urban" & session == 1 & run == 2)
    U_S2_R1 <- data %>% filter(task == "Urban" & session == 2 & run == 1)
    U_S2_R2 <- data %>% filter(task == "Urban" & session == 2 & run == 2)
    C_F1    <- data %>% filter(task == "Country" & session == 0 & run == 1)
    C_S1_R1 <- data %>% filter(task == "Country" & session == 1 & run == 1)
    C_S1_R2 <- data %>% filter(task == "Country" & session == 1 & run == 2)
    C_S2_R1 <- data %>% filter(task == "Country" & session == 2 & run == 1)
    C_S2_R2 <- data %>% filter(task == "Country" & session == 2 & run == 2)
    
    U_F1vsS1R1    <- ICC(data.frame(U_F1[,5],U_S1_R1[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(U_F1[,5],U_S1_R1[,5])
    colnames(totest) <- c("x","y")
    U_F1vsS1R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    U_F1vsS1R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    U_F1vsS1R1$CV_percent  <- sd(data.frame(rbind(U_F1[,5],U_S1_R1[,5]))[,1])/mean(data.frame(rbind(U_F1[,5],U_S1_R1[,5]))[,1])*100
    U_F1vsS1R1$t_test_t    <- t.test(data.frame(U_F1[,5],U_S1_R1[,5])[,1],data.frame(U_F1[,5],U_S1_R1[,5])[,2],paired = T)$statistic
    U_F1vsS1R1$t_test_df  <- t.test(data.frame(U_F1[,5],U_S1_R1[,5])[,1],data.frame(U_F1[,5],U_S1_R1[,5])[,2],paired = T)$parameter
    U_F1vsS1R1$t_test_p    <- t.test(data.frame(U_F1[,5],U_S1_R1[,5])[,1],data.frame(U_F1[,5],U_S1_R1[,5])[,2],paired = T)$p.value
    U_F1vsS1R1$mean_1      <- U_F1 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_F1vsS1R1$sd_1        <- U_F1 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_F1vsS1R1$mean_2      <- U_S1_R1 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_F1vsS1R1$sd_2        <- U_S1_R1 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(U_F1vsS1R1)  <-as.name("Fam VS S1R1")
    
    
    U_S1R1vsS1R2    <- ICC(data.frame(U_S1_R2[,5],U_S1_R1[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(U_S1_R2[,5],U_S1_R1[,5])
    colnames(totest) <- c("x","y")
    U_S1R1vsS1R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    U_S1R1vsS1R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    U_S1R1vsS1R2$CV_percent  <- sd(data.frame(rbind(U_S1_R2[,5],U_S1_R1[,5]))[,1])/mean(data.frame(rbind(U_S1_R2[,5],U_S1_R1[,5]))[,1])*100
    U_S1R1vsS1R2$t_test_t    <- t.test(data.frame(U_S1_R2[,5],U_S1_R1[,5])[,1],data.frame(U_S1_R2[,5],U_S1_R1[,5])[,2],paired = T)$statistic
    U_S1R1vsS1R2$t_test_df    <- t.test(data.frame(U_S1_R2[,5],U_S1_R1[,5])[,1],data.frame(U_S1_R2[,5],U_S1_R1[,5])[,2],paired = T)$parameter
    U_S1R1vsS1R2$t_test_p    <- t.test(data.frame(U_S1_R2[,5],U_S1_R1[,5])[,1],data.frame(U_S1_R2[,5],U_S1_R1[,5])[,2],paired = T)$p.value
    U_S1R1vsS1R2$mean_1      <- U_S1_R1 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R1vsS1R2$sd_1        <- U_S1_R1 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R1vsS1R2$mean_2      <- U_S1_R2 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R1vsS1R2$sd_2        <- U_S1_R2 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(U_S1R1vsS1R2)  <- as.name("Within day 1")
    
    U_S2R1vsS2R2    <- ICC(data.frame(U_S2_R1[,5],U_S2_R2[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(U_S2_R1[,5],U_S2_R2[,5])
    colnames(totest) <- c("x","y")
    U_S2R1vsS2R2$pearsons_r   <- cor.test(formula = ~x+y, data = totest)$estimate
    U_S2R1vsS2R2$pearsons_p   <- cor.test(formula = ~x+y, data = totest)$p.value
    U_S2R1vsS2R2$CV_percent   <- sd(data.frame(rbind(U_S2_R1[,5],U_S2_R2[,5]))[,1])/mean(data.frame(rbind(U_S2_R1[,5],U_S2_R2[,5]))[,1])*100
    U_S2R1vsS2R2$t_test_t     <- t.test(data.frame(U_S2_R1[,5],U_S2_R2[,5])[,1],data.frame(U_S2_R1[,5],U_S2_R2[,5])[,2],paired = T)$statistic
    U_S2R1vsS2R2$t_test_df    <- t.test(data.frame(U_S2_R1[,5],U_S2_R2[,5])[,1],data.frame(U_S2_R1[,5],U_S2_R2[,5])[,2],paired = T)$parameter
    U_S2R1vsS2R2$t_test_p     <- t.test(data.frame(U_S2_R1[,5],U_S2_R2[,5])[,1],data.frame(U_S2_R1[,5],U_S2_R2[,5])[,2],paired = T)$p.value
    U_S2R1vsS2R2$mean_1       <- U_S2_R1 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S2R1vsS2R2$sd_1         <- U_S2_R1 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S2R1vsS2R2$mean_2       <- U_S2_R2 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S2R1vsS2R2$sd_2         <- U_S2_R2 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(U_S2R1vsS2R2)   <- as.name("Within day 2")
    
    U_S1R1vsS2R1    <- ICC(data.frame(U_S2_R1[,5],U_S1_R1[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(U_S2_R1[,5],U_S1_R1[,5])
    colnames(totest) <- c("x","y")
    U_S1R1vsS2R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    U_S1R1vsS2R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    U_S1R1vsS2R1$CV_percent  <- sd(data.frame(rbind(U_S2_R1[,5],U_S1_R1[,5]))[,1])/mean(data.frame(rbind(U_S2_R1[,5],U_S1_R1[,5]))[,1])*100
    U_S1R1vsS2R1$t_test_t    <- t.test(data.frame(U_S2_R1[,5],U_S1_R1[,5])[,1],data.frame(U_S2_R1[,5],U_S1_R1[,5])[,2],paired = T)$statistic
    U_S1R1vsS2R1$t_test_df    <- t.test(data.frame(U_S2_R1[,5],U_S1_R1[,5])[,1],data.frame(U_S2_R1[,5],U_S1_R1[,5])[,2],paired = T)$parameter
    U_S1R1vsS2R1$t_test_p    <- t.test(data.frame(U_S2_R1[,5],U_S1_R1[,5])[,1],data.frame(U_S2_R1[,5],U_S1_R1[,5])[,2],paired = T)$p.value
    U_S1R1vsS2R1$mean_1      <- U_S1_R1 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R1vsS2R1$sd_1        <- U_S1_R1 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R1vsS2R1$mean_2      <- U_S2_R1 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R1vsS2R1$sd_2        <- U_S2_R1 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(U_S1R1vsS2R1)  <- as.name("Within run 1")
    
    U_S1R2vsS2R2    <- ICC(data.frame(U_S1_R2[,5],U_S2_R2[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(U_S1_R2[,5],U_S2_R2[,5])
    colnames(totest) <- c("x","y")
    U_S1R2vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    U_S1R2vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    U_S1R2vsS2R2$CV_percent  <- sd(data.frame(rbind(U_S1_R2[,5],U_S2_R2[,5]))[,1])/mean(data.frame(rbind(U_S1_R2[,5],U_S2_R2[,5]))[,1])*100
    U_S1R2vsS2R2$t_test_t    <- t.test(data.frame(U_S1_R2[,5],U_S2_R2[,5])[,1],data.frame(U_S1_R2[,5],U_S2_R2[,5])[,2],paired = T)$statistic
    U_S1R2vsS2R2$t_test_df    <- t.test(data.frame(U_S1_R2[,5],U_S2_R2[,5])[,1],data.frame(U_S1_R2[,5],U_S2_R2[,5])[,2],paired = T)$parameter
    U_S1R2vsS2R2$t_test_p    <- t.test(data.frame(U_S1_R2[,5],U_S2_R2[,5])[,1],data.frame(U_S1_R2[,5],U_S2_R2[,5])[,2],paired = T)$p.value
    U_S1R2vsS2R2$mean_1      <- U_S1_R2 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric() %>% round(3)
    U_S1R2vsS2R2$sd_1        <- U_S1_R2 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R2vsS2R2$mean_2      <- U_S2_R2 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    U_S1R2vsS2R2$sd_2        <- U_S2_R2 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(U_S1R2vsS2R2)  <- as.name("Within run 2")
    
    Urban_combined <- rbind(U_F1vsS1R1,U_S1R1vsS1R2,U_S2R1vsS2R2,U_S1R1vsS2R1,U_S1R2vsS2R2)
    Urban_combined <- Urban_combined %>% mutate(pearson = paste0("r = ",pearsons_r %>% round(3), ", p = ",round(pearsons_p,3)),
                                                    ICCsumm = paste0("ICC = ",round(ICC,3), ", p = ",round(p,3)),
                                                    ttest = paste0("t = ",round(t_test_t,3), ", df = ",round(t_test_df,3), ", p = ",round(t_test_p,3))) %>%
                                             select(mean_1,sd_1,mean_2,sd_2,pearson,ICCsumm,CV_percent,ttest)
    row.names(Urban_combined) <- c(as.name("Fam VS S1R1"),as.name("Within day 1"),as.name("Within day 2"),as.name("Within run 1"),as.name("Within run 2"))


    
    table2office(Urban_combined,paste0("G:\\Shared drives\\Projects\\2018 Driving Sim Reproducibility\\ICC\\Driving\\",
                                         "Urban_",data_name,variable,"_ICC.docx"), type = "DOC", pointsize = 9, append = FALSE, add.rownames = TRUE )
 
    
    C_F1vsS1R1    <- ICC(data.frame(C_F1[,5],C_S1_R1[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(C_F1[,5],C_S1_R1[,5])
    colnames(totest) <- c("x","y")
    C_F1vsS1R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_F1vsS1R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_F1vsS1R1$CV_percent  <- sd(data.frame(rbind(C_F1[,5],C_S1_R1[,5]))[,1])/mean(data.frame(rbind(C_F1[,5],C_S1_R1[,5]))[,1])*100
    C_F1vsS1R1$t_test_t    <- t.test(data.frame(C_F1[,5],C_S1_R1[,5])[,1],data.frame(C_F1[,5],C_S1_R1[,5])[,2],paired = T)$statistic
    C_F1vsS1R1$t_test_df    <- t.test(data.frame(C_F1[,5],C_S1_R1[,5])[,1],data.frame(C_F1[,5],C_S1_R1[,5])[,2],paired = T)$parameter
    C_F1vsS1R1$t_test_p    <- t.test(data.frame(C_F1[,5],C_S1_R1[,5])[,1],data.frame(C_F1[,5],C_S1_R1[,5])[,2],paired = T)$p.value
    C_F1vsS1R1$mean_1      <- C_F1 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_F1vsS1R1$sd_1        <- C_F1 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_F1vsS1R1$mean_2      <- C_S1_R1 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_F1vsS1R1$sd_2        <- C_S1_R1 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(C_F1vsS1R1)  <-"Fam VS S1R1"
    
    C_S1R1vsS1R2    <- ICC(data.frame(C_S1_R2[,5],C_S1_R1[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(C_S1_R2[,5],C_S1_R1[,5])
    colnames(totest) <- c("x","y")
    C_S1R1vsS1R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS1R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS1R2$CV_percent  <- sd(data.frame(rbind(C_S1_R2[,5],C_S1_R1[,5]))[,1])/mean(data.frame(rbind(C_S1_R2[,5],C_S1_R1[,5]))[,1])*100
    C_S1R1vsS1R2$t_test_t    <- t.test(data.frame(C_S1_R2[,5],C_S1_R1[,5])[,1],data.frame(C_S1_R2[,5],C_S1_R1[,5])[,2],paired = T)$statistic
    C_S1R1vsS1R2$t_test_df    <- t.test(data.frame(C_S1_R2[,5],C_S1_R1[,5])[,1],data.frame(C_S1_R2[,5],C_S1_R1[,5])[,2],paired = T)$parameter
    C_S1R1vsS1R2$t_test_p    <- t.test(data.frame(C_S1_R2[,5],C_S1_R1[,5])[,1],data.frame(C_S1_R2[,5],C_S1_R1[,5])[,2],paired = T)$p.value
    C_S1R1vsS1R2$mean_1      <- C_S1_R1 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_S1R1vsS1R2$sd_1        <- C_S1_R1 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_S1R1vsS1R2$mean_2      <- C_S1_R2 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_S1R1vsS1R2$sd_2        <- C_S1_R2 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(C_S1R1vsS1R2)  <- "Within day 1"

    C_S2R1vsS2R2    <- ICC(data.frame(C_S2_R1[,5],C_S2_R2[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(C_S2_R1[,5],C_S2_R2[,5])
    colnames(totest) <- c("x","y")
    C_S2R1vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S2R1vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S2R1vsS2R2$CV_percent  <- sd(data.frame(rbind(C_S2_R1[,5],C_S2_R2[,5]))[,1])/mean(data.frame(rbind(C_S2_R1[,5],C_S2_R2[,5]))[,1])*100
    C_S2R1vsS2R2$t_test_t    <- t.test(data.frame(C_S2_R1[,5],C_S2_R2[,5])[,1],data.frame(C_S2_R1[,5],C_S2_R2[,5])[,2],paired = T)$statistic
    C_S2R1vsS2R2$t_test_df    <- t.test(data.frame(C_S2_R1[,5],C_S2_R2[,5])[,1],data.frame(C_S2_R1[,5],C_S2_R2[,5])[,2],paired = T)$parameter
    C_S2R1vsS2R2$t_test_p    <- t.test(data.frame(C_S2_R1[,5],C_S2_R2[,5])[,1],data.frame(C_S2_R1[,5],C_S2_R2[,5])[,2],paired = T)$p.value
    C_S2R1vsS2R2$mean_1      <- C_S2_R1 %>% summarise(mean_1 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_S2R1vsS2R2$sd_1        <- C_S2_R1 %>% summarise(sd_1 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_S2R1vsS2R2$mean_2      <- C_S2_R2 %>% summarise(mean_2 = mean(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    C_S2R1vsS2R2$sd_2        <- C_S2_R2 %>% summarise(sd_2 = sd(eval(parse(text = variable)))) %>% as.numeric()%>% round(3)
    row.names(C_S2R1vsS2R2)  <- "Within day 2"

    C_S1R1vsS2R1    <- ICC(data.frame(C_S2_R1[,5],C_S1_R1[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(C_S2_R1[,5],C_S1_R1[,5])
    colnames(totest) <- c("x","y")
    C_S1R1vsS2R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS2R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS2R1$CV_percent  <- sd(data.frame(rbind(C_S2_R1[,5],C_S1_R1[,5]))[,1])/mean(data.frame(rbind(C_S2_R1[,5],C_S1_R1[,5]))[,1])*100
    C_S1R1vsS2R1$t_test_t    <- t.test(data.frame(C_S2_R1[,5],C_S1_R1[,5])[,1],data.frame(C_S2_R1[,5],C_S1_R1[,5])[,2],paired = T)$statistic
    C_S1R1vsS2R1$t_test_df    <- t.test(data.frame(C_S2_R1[,5],C_S1_R1[,5])[,1],data.frame(C_S2_R1[,5],C_S1_R1[,5])[,2],paired = T)$parameter
    C_S1R1vsS2R1$t_test_p    <- t.test(data.frame(C_S2_R1[,5],C_S1_R1[,5])[,1],data.frame(C_S2_R1[,5],C_S1_R1[,5])[,2],paired = T)$p.value
    C_S1R1vsS2R1$mean_1      <- C_S1_R1 %>% summarise(mean_1 = mean(eval(parse(text = variable))))%>% round(3)
    C_S1R1vsS2R1$sd_1        <- C_S1_R1 %>% summarise(sd_1 = sd(eval(parse(text = variable))))%>% round(3)
    C_S1R1vsS2R1$mean_2      <- C_S2_R1 %>% summarise(mean_2 = mean(eval(parse(text = variable))))%>% round(3)
    C_S1R1vsS2R1$sd_2        <- C_S2_R1 %>% summarise(sd_2 = sd(eval(parse(text = variable))))%>% round(3)
    row.names(C_S1R1vsS2R1)  <- "Within run 1"

    C_S1R2vsS2R2    <- ICC(data.frame(C_S1_R2[,5],C_S2_R2[,5]))$results[3,c(1,2,3,6)]
    totest <-data.frame(C_S1_R2[,5],C_S2_R2[,5])
    colnames(totest) <- c("x","y")
    C_S1R2vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R2vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R2vsS2R2$CV_percent  <- sd(data.frame(rbind(C_S1_R2[,5],C_S2_R2[,5]))[,1])/mean(data.frame(rbind(C_S1_R2[,5],C_S2_R2[,5]))[,1])*100
    C_S1R2vsS2R2$t_test_t    <- t.test(data.frame(C_S1_R2[,5],C_S2_R2[,5])[,1],data.frame(C_S1_R2[,5],C_S2_R2[,5])[,2],paired = T)$statistic
    C_S1R2vsS2R2$t_test_df    <- t.test(data.frame(C_S1_R2[,5],C_S2_R2[,5])[,1],data.frame(C_S2_R1[,5],C_S1_R1[,5])[,2],paired = T)$parameter
    C_S1R2vsS2R2$t_test_p    <- t.test(data.frame(C_S1_R2[,5],C_S2_R2[,5])[,1],data.frame(C_S1_R2[,5],C_S2_R2[,5])[,2],paired = T)$p.value
    C_S1R2vsS2R2$mean_1      <- C_S1_R2 %>% summarise(mean_1 = mean(eval(parse(text = variable))))%>% round(3)
    C_S1R2vsS2R2$sd_1        <- C_S1_R2 %>% summarise(sd_1 = sd(eval(parse(text = variable))))%>% round(3)
    C_S1R2vsS2R2$mean_2      <- C_S2_R2 %>% summarise(mean_2 = mean(eval(parse(text = variable))))%>% round(3)
    C_S1R2vsS2R2$sd_2        <- C_S2_R2 %>% summarise(sd_2 = sd(eval(parse(text = variable))))%>% round(3)
    row.names(C_S1R2vsS2R2)  <- "Within run 2"
    
    Country_combined <- rbind(C_F1vsS1R1,C_S1R1vsS1R2,C_S2R1vsS2R2,C_S1R1vsS2R1,C_S1R2vsS2R2)
    Country_combined <- Country_combined %>% mutate(pearson = paste0("r = ",pearsons_r %>% round(3), ", p = ",round(pearsons_p,3)),
                                                    ICCsumm = paste0("ICC = ",round(ICC,3), ", p = ",round(p,3)),
                                                    ttest = paste0("t = ",round(t_test_t,3), ", df = ",round(t_test_df,3), ", p = ",round(t_test_p,3))) %>%
                                             select(mean_1,sd_1,mean_2,sd_2,pearson,ICCsumm,CV_percent,ttest)
    row.names(Country_combined) <- c(as.name("Fam VS S1R1"),as.name("Within day 1"),as.name("Within day 2"),as.name("Within run 1"),as.name("Within run 2"))


    
    table2office(Country_combined,paste0("G:\\Shared drives\\Projects\\2018 Driving Sim Reproducibility\\ICC\\Driving\\",
                                         "Country_",data_name,"_",variable,"_ICC.docx"), type = "DOC", pointsize = 9, append = FALSE, add.rownames = TRUE )

}
```



```{r}
create_ICC_dataframe_driving(summarised_driving_data, 'avg_speed',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'sd_speed',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'avg_throttle',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'sd_throttle',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'avg_lanepos',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'sd_lanepos',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'avg_steering',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'sd_steering',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'avg_filt_lanepos',data_name = "driving")
create_ICC_dataframe_driving(summarised_driving_data, 'sd_filt_lanepos',data_name = "driving")

# avg_speed, sd_speed, avg_throttle, sd_throttle, avg_lanepos, sd_lanepos, avg_steering, sd_steering, ang_filt_lanepos, sd_filt_lanepos
# 

```


```{r}

create_ICC_dataframe_saccades <- function(database,variable,data_name){
    data <- database %>% select(ID,session = Session,run = Run,variable)
    C_F1    <- data %>% filter(session == "F" & run == 1)
    C_S1_R1 <- data %>% filter(session == 1 & run == 1)
    C_S1_R2 <- data %>% filter(session == 1 & run == 2)
    C_S2_R1 <- data %>% filter(session == 2 & run == 1)
    C_S2_R2 <- data %>% filter(session == 2 & run == 2)
    
    combined_data          <- C_F1 %>% full_join(C_S1_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1            <- combined_data[,4]
    timepoint_2            <- combined_data[,7]
    C_F1vsS1R1             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                 <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)       <- c("x","y")
    C_F1vsS1R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_F1vsS1R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_F1vsS1R1$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_F1vsS1R1$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_F1vsS1R1$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_F1vsS1R1$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_F1vsS1R1$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_F1vsS1R1)  <- "Fam VS S1R1"
    
    combined_data            <- C_S1_R1 %>% full_join(C_S1_R2, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S1R1vsS1R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R1vsS1R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS1R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS1R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R1vsS1R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R1vsS1R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R1vsS1R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R1vsS1R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R1vsS1R2)  <- "Within day 1"

    combined_data            <- C_S2_R1 %>% full_join(C_S2_R2, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S2R1vsS2R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S2R1vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S2R1vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S2R1vsS2R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S2R1vsS2R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S2R1vsS2R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S2R1vsS2R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S2R1vsS2R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S2R1vsS2R2)  <- "Within day 2"

    combined_data            <- C_S1_R1 %>% full_join(C_S2_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S1R1vsS2R1             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R1vsS2R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS2R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS2R1$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R1vsS2R1$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R1vsS2R1$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R1vsS2R1$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R1vsS2R1$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R1vsS2R1)  <- "Within run 1"

    combined_data            <- C_S1_R2 %>% full_join(C_S2_R2, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S1R2vsS2R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R2vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R2vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R2vsS2R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R2vsS2R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R2vsS2R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R2vsS2R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R2vsS2R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R2vsS2R2)  <- "Within run 2"
    
    Country_combined <- rbind(C_F1vsS1R1,C_S1R1vsS1R2,C_S2R1vsS2R2,C_S1R1vsS2R1,C_S1R2vsS2R2)
    Country_combined <- Country_combined %>% mutate(pearson = paste0("r = ",pearsons_r %>% round(3), ", p = ",round(pearsons_p,3)),
                                                    ICCsumm = paste0("ICC = ",round(ICC,3), ", p = ",round(p,3)),
                                                    ttest = paste0("t = ",round(t_test_t,3), ", df = ",round(t_test_df,3), ", p = ",round(t_test_p,3))) %>%
                                             select(mean_1,sd_1,mean_2,sd_2,pearson,ICCsumm,CV_percent,ttest)
    row.names(Country_combined) <- c(as.name("Fam VS S1R1"),as.name("Within day 1"),as.name("Within day 2"),as.name("Within run 1"),as.name("Within run 2"))


    
    table2office(Country_combined,paste0("G:\\Shared drives\\Projects\\2018 Driving Sim Reproducibility\\ICC\\Saccades\\",
                                         "Country_",data_name,"_",variable,"_ICC.docx"), type = "DOC", pointsize = 9, append = FALSE, add.rownames = TRUE )
   
}

```

# Analysis of Saccade Data
```{r}
saccades_database <- read_csv("G:\\Shared drives\\Projects\\2018 Driving Sim Reproducibility\\4_Analysis_Saccades\\Individual_saccade_database.csv")

attention_performance_database <- read_csv("G:/Shared drives/Projects\\2018 Driving Sim Reproducibility\\task_response_database.csv") %>%
  select(ID, Session, Run, `trial number`, task_condition, task_arrow_dir, stimulus_left_right, congruency, target_deviation, response_type,
                   reaction_time)

saccade_attn_joined <- saccades_database %>% left_join(attention_performance_database, 
                                  by = c("ID","Session","Run","trial number")) %>% 
                        select(ID,Session,Run,`trial number`,condition, task_condition, eye, 
                               stimulus_left_right, task_arrow_dir, congruency,target_deviation,
                               `correct direction`, duration, `dva/s`, `peak_dva/s`, amplitude, saccade_size, `stimulus onset delay`,
                               reaction_time, response_type) 

saccade_attn_joined <- saccade_attn_joined %>% mutate(Timepoint = factor(str_c(Session,Run, sep = "_")))
names(saccade_attn_joined)

organised_saccade_data <- saccade_attn_joined %>% dplyr::filter(amplitude > 2) %>% 
                                                  filter(is.double(`peak_dva/s`)) %>% 
                                                  filter(`peak_dva/s` < 1500) %>% 
                                                  filter(`dva/s` < 1000)
organised_saccade_data$saccade_bin <- if_else((organised_saccade_data$amplitude > 5) & (organised_saccade_data$amplitude < 15), "5-15",
                                         if_else((organised_saccade_data$amplitude > 20) & (organised_saccade_data$amplitude < 30), "20-30",
                                         if_else((organised_saccade_data$amplitude > 30) & (organised_saccade_data$amplitude < 40), "30-40","NA"))) %>%
                                         as.factor()

saccade_by_bin <- organised_saccade_data %>% group_by(ID,Session,Run, saccade_bin) %>% 
                                             filter(saccade_bin != "NA") %>%
                                             summarise(mean_pv = mean(`peak_dva/s`), 
                                                       mean_dur = mean(duration), 
                                                       mean_avgv = mean(`dva/s`), 
                                                       n = n())

saccades_5_to_15  <-  saccade_by_bin %>% filter(saccade_bin == "5-15") %>% filter(n >= 5) %>% select(-n)
saccades_20_to_30 <-  saccade_by_bin %>% filter(saccade_bin == "20-30") %>% filter(n >= 5)%>% select(-n)
saccades_30_to_40 <-  saccade_by_bin %>% filter(saccade_bin == "30-40") %>% filter(n >= 5)%>% select(-n)

```

```{r}
create_ICC_dataframe_saccades(saccades_5_to_15,'mean_pv',"n_5_peak_velocity 5-15 degrees")
create_ICC_dataframe_saccades(saccades_20_to_30,'mean_pv',"n_5_peak_velocity 20-30 degrees")
create_ICC_dataframe_saccades(saccades_30_to_40,'mean_pv',"n_5_peak_velocity 30-40 degrees")

create_ICC_dataframe_saccades(saccades_5_to_15,'mean_dur',"n_5_saccade_duration 5-15 degrees")
create_ICC_dataframe_saccades(saccades_20_to_30,'mean_dur',"n_5_saccade_duration 20-30 degrees")
create_ICC_dataframe_saccades(saccades_30_to_40,'mean_dur',"n_5_saccade_duration 30-40 degrees")

create_ICC_dataframe_saccades(saccades_5_to_15,'mean_avgv',"n_5_mean_velocity 5-15 degrees")
create_ICC_dataframe_saccades(saccades_20_to_30,'mean_avgv',"n_5_mean_velocity 20-30 degrees")
create_ICC_dataframe_saccades(saccades_30_to_40,'mean_avgv',"n_5_mean_velocity 30-40 degrees")
```

# Analysis of Task Performance Data
```{r}
task_response_performance <- read_csv("G:\\Shared drives\\Projects\\2018 Driving Sim Reproducibility\\task_response_database.csv")
task_response_performance_summarised_wo_cong_accuracy <- task_response_performance %>% 
                                        group_by(ID,Session,Run,target_deviation) %>% 
                                        summarise(count = n(),
                                                  percent_correct = mean(response_type == "Correct", na.rm = TRUE), 
                                                  percent_miss = mean(response_type == "Miss", na.rm = TRUE),
                                                  percent_incorrect = mean(response_type == "Incorrect", na.rm = TRUE))

task_response_performance_summarised_wo_cong_RT <- task_response_performance %>% 
                                        group_by(ID,Session,Run,target_deviation) %>% 
                                        filter(response_type == "Correct") %>%
                                        summarise(mean_RT = mean(reaction_time))

task_response_performance_summarised_w_cong_RT <- task_response_performance %>% 
                                        group_by(ID,Session,Run,congruency,target_deviation) %>% 
                                        filter(response_type == "Correct") %>%
                                        summarise(mean_RT = mean(reaction_time))

task_response_performance_summarised_w_cong_accuracy <- task_response_performance %>% 
                                        group_by(ID,Session,Run,congruency,target_deviation) %>% 
                                        summarise(percent_correct = mean(response_type == "Correct", na.rm = TRUE), 
                                                  percent_miss = mean(response_type == "Miss", na.rm = TRUE),
                                                  percent_incorrect = mean(response_type == "Incorrect", na.rm = TRUE))

task_response_performance_summarised_w_cong <- task_response_performance_summarised_w_cong_RT %>% left_join(task_response_performance_summarised_w_cong_accuracy)
task_response_performance_summarised_wo_cong <- task_response_performance_summarised_wo_cong_RT %>% left_join(task_response_performance_summarised_wo_cong_accuracy)

task_response_performance_summarised_cong10    <- task_response_performance_summarised_w_cong %>% filter(target_deviation == 10, congruency == "Congruent")
task_response_performance_summarised_incong10  <- task_response_performance_summarised_w_cong %>% filter(target_deviation == 10, congruency == "Incongruent")
task_response_performance_summarised_cong25   <- task_response_performance_summarised_w_cong %>% filter(target_deviation == 25, congruency == "Congruent")
task_response_performance_summarised_incong25  <- task_response_performance_summarised_w_cong %>% filter(target_deviation == 25, congruency == "Incongruent")
task_response_performance_summarised_cong35   <- task_response_performance_summarised_w_cong %>% filter(target_deviation == 35, congruency == "Congruent")
task_response_performance_summarised_incong35  <- task_response_performance_summarised_w_cong %>% filter(target_deviation == 35, congruency == "Incongruent")



task_response_performance_summarised10 <- task_response_performance_summarised_wo_cong %>% filter(target_deviation == 10)
task_response_performance_summarised25 <- task_response_performance_summarised_wo_cong %>% filter(target_deviation == 25)
task_response_performance_summarised35 <- task_response_performance_summarised_wo_cong %>% filter(target_deviation == 35)
```


```{r}
create_ICC_dataframe_attention <- function(database,variable,data_name){
    if ('congruency' %in% names(database)) {
    data <- database %>% select(ID,session = Session,run = Run,variable,-congruency)
    C_F1    <- data %>% filter(session == "F" & run == 1)
    C_S1_R1 <- data %>% filter(session == 1 & run == 1)
    C_S1_R2 <- data %>% filter(session == 1 & run == 2)
    C_S2_R1 <- data %>% filter(session == 2 & run == 1)
    C_S2_R2 <- data %>% filter(session == 2 & run == 2)
    
    combined_data          <- C_F1 %>% full_join(C_S1_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1            <- combined_data[,5]
    timepoint_2            <- combined_data[,9]
    C_F1vsS1R1             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                 <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)       <- c("x","y")
    C_F1vsS1R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_F1vsS1R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_F1vsS1R1$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_F1vsS1R1$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_F1vsS1R1$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_F1vsS1R1$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_F1vsS1R1$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_F1vsS1R1)  <- "Fam VS S1R1"
    
    combined_data            <- C_F1 %>% full_join(C_S1_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,5]
    timepoint_2              <- combined_data[,9]
    C_S1R1vsS1R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R1vsS1R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS1R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS1R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R1vsS1R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R1vsS1R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R1vsS1R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R1vsS1R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R1vsS1R2)  <- "Within day 1"

    combined_data            <- C_S2_R1 %>% full_join(C_S2_R2, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,5]
    timepoint_2              <- combined_data[,9]
    C_S2R1vsS2R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S2R1vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S2R1vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S2R1vsS2R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S2R1vsS2R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S2R1vsS2R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S2R1vsS2R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S2R1vsS2R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S2R1vsS2R2)  <- "Within day 2"

    combined_data            <- C_S1_R1 %>% full_join(C_S2_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,5]
    timepoint_2              <- combined_data[,9]
    C_S1R1vsS2R1             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R1vsS2R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS2R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS2R1$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R1vsS2R1$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R1vsS2R1$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R1vsS2R1$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R1vsS2R1$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R1vsS2R1)  <- "Within run 1"

    combined_data            <- C_S1_R2 %>% full_join(C_S2_R2, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,5]
    timepoint_2              <- combined_data[,9]
    C_S1R2vsS2R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R2vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R2vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R2vsS2R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R2vsS2R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R2vsS2R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R2vsS2R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R2vsS2R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R2vsS2R2)  <- "Within run 2"
    } else {
    data <- database %>% select(ID,session = Session,run = Run,variable)
    C_F1    <- data %>% filter(session == "F" & run == 1)
    C_S1_R1 <- data %>% filter(session == 1 & run == 1)
    C_S1_R2 <- data %>% filter(session == 1 & run == 2)
    C_S2_R1 <- data %>% filter(session == 2 & run == 1)
    C_S2_R2 <- data %>% filter(session == 2 & run == 2)
    
    combined_data          <- C_F1 %>% full_join(C_S1_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1            <- combined_data[,4]
    timepoint_2            <- combined_data[,7]
    C_F1vsS1R1             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                 <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)       <- c("x","y")
    C_F1vsS1R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_F1vsS1R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_F1vsS1R1$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_F1vsS1R1$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_F1vsS1R1$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_F1vsS1R1$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_F1vsS1R1$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_F1vsS1R1$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_F1vsS1R1)  <- "Fam VS S1R1"
    
    combined_data            <- C_F1 %>% full_join(C_S1_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S1R1vsS1R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R1vsS1R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS1R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS1R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R1vsS1R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R1vsS1R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R1vsS1R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R1vsS1R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R1vsS1R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R1vsS1R2)  <- "Within day 1"

    combined_data            <- C_S2_R1 %>% full_join(C_S2_R2, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S2R1vsS2R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S2R1vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S2R1vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S2R1vsS2R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S2R1vsS2R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S2R1vsS2R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S2R1vsS2R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S2R1vsS2R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S2R1vsS2R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S2R1vsS2R2)  <- "Within day 2"

    combined_data            <- C_S1_R1 %>% full_join(C_S2_R1, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S1R1vsS2R1             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R1vsS2R1$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R1vsS2R1$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R1vsS2R1$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R1vsS2R1$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R1vsS2R1$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R1vsS2R1$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R1vsS2R1$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R1vsS2R1$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R1vsS2R1)  <- "Within run 1"

    combined_data            <- C_S1_R2 %>% full_join(C_S2_R2, by = "ID") %>% drop_na() %>% data.frame()
    timepoint_1              <- combined_data[,4]
    timepoint_2              <- combined_data[,7]
    C_S1R2vsS2R2             <- ICC(data.frame(timepoint_1,timepoint_2))$results[3,c(1,2,3,6)]
    totest                   <- data.frame(timepoint_1, timepoint_2)
    colnames(totest)         <- c("x","y")
    C_S1R2vsS2R2$pearsons_r  <- cor.test(formula = ~x+y, data = totest)$estimate
    C_S1R2vsS2R2$pearsons_p  <- cor.test(formula = ~x+y, data = totest)$p.value
    C_S1R2vsS2R2$CV_percent  <- sd(data.frame(rbind(timepoint_1,timepoint_2))[,1])/mean(data.frame(rbind(timepoint_1,timepoint_2))[,1])*100
    C_S1R2vsS2R2$t_test_t    <- t.test(timepoint_1,timepoint_2,paired = T)$statistic
    C_S1R2vsS2R2$t_test_df   <- t.test(timepoint_1,timepoint_2,paired = T)$parameter
    C_S1R2vsS2R2$t_test_p    <- t.test(timepoint_1,timepoint_2,paired = T)$p.value
    C_S1R2vsS2R2$mean_1      <- mean(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$sd_1        <- sd(timepoint_1) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$mean_2      <- mean(timepoint_2) %>% as.numeric() %>% round(3)
    C_S1R2vsS2R2$sd_2        <- sd(timepoint_2) %>% as.numeric() %>% round(3)
    row.names(C_S1R2vsS2R2)  <- "Within run 2" 
    }
    
    Country_combined <- rbind(C_F1vsS1R1,C_S1R1vsS1R2,C_S2R1vsS2R2,C_S1R1vsS2R1,C_S1R2vsS2R2)
    Country_combined <- Country_combined %>% mutate(pearson = paste0("r = ",pearsons_r %>% round(3), ", p = ",round(pearsons_p,3)),
                                                    ICCsumm = paste0("ICC = ",round(ICC,3), ", p = ",round(p,3)),
                                                    ttest = paste0("t = ",round(t_test_t,3), ", df = ",round(t_test_df,3), ", p = ",round(t_test_p,3))) %>%
                                             select(mean_1,sd_1,mean_2,sd_2,pearson,ICCsumm,CV_percent,ttest)
    row.names(Country_combined) <- c(as.name("Fam VS S1R1"),as.name("Within day 1"),as.name("Within day 2"),as.name("Within run 1"),as.name("Within run 2"))


    
    table2office(Country_combined,paste0("G:\\Shared drives\\Projects\\2018 Driving Sim Reproducibility\\ICC\\Task Performance\\",
                                         "Country_",data_name,"_",variable,"_ICC.docx"), type = "DOC", pointsize = 9, append = FALSE, add.rownames = TRUE )

}

```


```{r}

create_ICC_dataframe_attention(task_response_performance_summarised_cong10, 'mean_RT',data_name = "congruent_10_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_incong10, 'mean_RT',data_name = "incongruent_10_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_cong25, 'mean_RT',data_name = "congruent_25_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_incong25, 'mean_RT',data_name = "incongruent_25_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_cong35, 'mean_RT',data_name = "congruent_35_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_incong35, 'mean_RT',data_name = "incongruent_35_task_performance")

create_ICC_dataframe_attention(task_response_performance_summarised_cong10, 'percent_correct',data_name = "congruent_10_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_incong10, 'percent_correct',data_name = "incongruent_10_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_cong25, 'percent_correct',data_name = "congruent_25_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_incong25, 'percent_correct',data_name = "incongruent_25_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_cong35, 'percent_correct',data_name = "congruent_35_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised_incong35, 'percent_correct',data_name = "incongruent_35_task_performance")


create_ICC_dataframe_attention(task_response_performance_summarised10, 'mean_RT',data_name = "combined_10_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised25, 'mean_RT',data_name = "combined_25_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised35, 'mean_RT',data_name = "combined_35_task_performance")

create_ICC_dataframe_attention(task_response_performance_summarised10, 'percent_correct',data_name = "combined_10_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised25, 'percent_correct',data_name = "combined_25_task_performance")
create_ICC_dataframe_attention(task_response_performance_summarised35, 'percent_correct',data_name = "combined_35_task_performance")

```
